Reduce prop drilling by passing children
Per a YouTube video,
make components children,
instead of passing props through them.
