+++ js/src/edit-block/components/side.js
@@ -12,19 +12,13 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { BlockPanel, FieldPanel } from './';
 import { BLOCK_PANEL, FIELD_PANEL } from '../constants';
 
 /**
  * @typedef {Object} SideProps The component props.
- * @property {import('./editor').CurrentLocation} currentLocation The currently selected location.
- * @property {import('./editor').IsNewField} isNewField Whether there is a new field.
+ * @property {React.ReactElement} children The component children.
  * @property {import('./editor').PanelDisplaying} panelDisplaying The panel currently displaying in the side, like 'block'.
- * @property {import('./editor').SelectedField|import('../constants').NoFieldSelected} selectedField The name of the selected field, if any.
- * @property {import('./editor').SetCurrentLocation} setCurrentLocation Sets the current location, like 'editor'.
- * @property {import('./editor').SetIsNewField} setIsNewField Sets if there is a new field.
  * @property {import('./editor').SetPanelDisplaying} setPanelDisplaying Sets the panel currently displaying in the side.
- * @property {import('./editor').SetSelectedField} setSelectedField Sets the selected field name.
  */
 
 /**
@@ -34,14 +28,9 @@ import { BLOCK_PANEL, FIELD_PANEL } from '../constants';
  * @return {React.ReactElement} The side area.
  */
 const Side = ( {
-	currentLocation,
-	isNewField,
+	children,
 	panelDisplaying,
-	selectedField,
-	setCurrentLocation,
-	setIsNewField,
 	setPanelDisplaying,
-	setSelectedField,
 } ) => {
 	const buttonClass = 'flex items-center h-12 px-5 text-sm focus:outline-none';
 
@@ -69,19 +58,7 @@ const Side = ( {
 					{ __( 'Field', 'genesis-custom-blocks' ) }
 				</button>
 			</div>
-			{ BLOCK_PANEL === panelDisplaying
-				? <BlockPanel />
-				: (
-					<FieldPanel
-						currentLocation={ currentLocation }
-						isNewField={ isNewField }
-						selectedField={ selectedField }
-						setCurrentLocation={ setCurrentLocation }
-						setIsNewField={ setIsNewField }
-						setSelectedField={ setSelectedField }
-					/>
-				)
-			}
+			{ children }
 		</div>
 	);
 };
